+++ src/CleanRegex/Internal/Prepared/Parser/Parser.php
@@ -6,7 +6,7 @@ use TRegx\CleanRegex\Internal\Prepared\Quoteable\Quoteable;
 
 interface Parser
 {
-    public function getDelimiterable(): string;
-
     public function parse(string $delimiter, QuotableFactory $quotableFactory): Quoteable;
+
+    public function getDelimiterable(): string;
 }