+++ ChangeLog.md
@@ -19,6 +19,7 @@ Incoming in 0.9.14
       or nonexistent group as well.
 
 * Features
+    * Add `Pattern::format()` #79
     * Add `Detail.textByteLength()` #88
     * Add `DetailGroup.textByteLength()` #88
     * Add `match()->flatMapAssoc()` #88